# runtime/packages/client/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Scrapyomama Runtime API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /
# Par défaut : session d'interface (cookie) ou clé d'API (Bearer, scopes de 13 § 8). Les routes publiques posent `security: []` ;
# les routes réservées à la session (registre `auth: 'session'`) posent `security: [{sessionCookie: []}]` : jamais une clé d'API
# sur une route d'administration (13 § 8). Le test de dérive compare ce mode d'authentification au registre du serveur.
security:
  - sessionCookie: []
  - apiKey: []
...
paths:
...
  /api/runs:
    get:
      operationId: listRuns
      tags: [runs]
      summary: « Tous les runs » et historique d'une API, filtrables, pagination serveur (06 § 2) ; extension de 05 § 4.2
      parameters:
        - name: api
          in: query
          description: Slug de l'API.
          schema:
            type: string
        - name: state
          in: query
          schema:
            $ref: '#/components/schemas/RunState'
        - name: trigger
          in: query
          schema:
            $ref: '#/components/schemas/RunTrigger'
        - name: since
          in: query
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Runs (métadonnées seulement, jamais d'items).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunList'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /api/runs/{id}:
    get:
      operationId: getRun
      tags: [runs]
      summary: Un run (= get_run) ; l'admin n'obtient que les métadonnées d'un run avec session d'autrui
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Run, avec sa cascade d'essais.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/runs/{id}/cancel:
    post:
      operationId: cancelRun
      tags: [runs]
      summary: Arrête un run ou une enquête (= cancel_run), essais et coûts engagés conservés
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Run annulé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunCancelled'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/runs/{id}/pause:
    post:
      # Extension de 05 § 4.2 (tâche 3.5) : 06 § 2 offre « Pause » et « Arrêter » sur une enquête ; 05 ne nomme que cancel et resume.
      operationId: pauseRun
      tags: [runs]
      summary: Met en pause une enquête ou un run en cours (action de l'utilisateur) ; essais et coûts engagés conservés, reprise par `resume`
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '202':
          $ref: '#/components/responses/Accepted'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/runs/{id}/resume:
    post:
      operationId: resumeRun
      tags: [runs]
      summary: Reprend un run en pause (action de l'utilisateur ; jamais en réponse à une vérification)
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '202':
          $ref: '#/components/responses/Accepted'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
...
  /api/runs/{id}/logs:
    get:
      operationId: listRunLogs
      tags: [runs]
      summary: Journal d'un run (`run_logs`, vue Technique de 06 § 2) ; extension de 05 § 4.2
      parameters:
        - $ref: '#/components/parameters/Id'
        - name: after
          in: query
          schema:
            type: integer
            minimum: 0
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Lignes du journal, en ordre croissant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunLogList'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-sy.session
      description: Session d'interface (13 § 5) ; `sy.session` sans le préfixe `__Host-` quand PUBLIC_URL est en http (développement).
    apiKey:
      type: http
      scheme: bearer
      description: Clé d'API `sy_live_…` à scopes (13 § 8) ; jamais de scope d'administration.
    deviceToken:
      type: http
      scheme: bearer
      description: Jeton d'appareil de l'extension (07 § 1), lié à (utilisateur, appareil) ; jamais accepté hors de /api/extension.
    metricsToken:
      type: http
      scheme: bearer
      description: METRICS_TOKEN (14 § 3), distinct des clés d'API.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      description: Curseur opaque renvoyé par la page précédente (`next_cursor`).
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
...
  responses:
    Error:
      description: "Erreur au format commun `{ error: { code, message } }` (le code est stable, jamais localisé)."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Accepted:
      description: Accepté ; run (ou enquête) à suivre par `GET /api/runs/{id}` ou le flux SSE (05 § 4.3).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RunAccepted'
...
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    Execution:
      description: Niveaux E1 à E6, du moins cher au plus cher (04 § 3.1).
      type: string
      enum: [fetch, fetch_in_page, playwright, agent_fetch, hybrid, agent]
    Network:
      type: string
      enum: [direct, dc_proxy, res_proxy, tunnel]
...
    RunTrigger:
      type: string
      enum: [mcp, rest, schedule, ui, canary]
    RunState:
      type: string
      enum: [queued, running, waiting_tunnel, succeeded, failed, cancelled, skipped_tunnel_offline, skipped_window, skipped_quota, skipped_status, skipped_overlap]
    RunOutcome:
      type: string
      enum: [clean, degraded, failed]
    FailureClass:
      description: Classe d'échec fermée (04b § 1), ou famille `llm_*` (08 § 1). Code stable, jamais localisé.
      type: string
      pattern: '^(transient|network|rate_limited|forbidden|blocked_by_protection|robots_disallowed|robots_unreachable|payment_required|auth_required|account_limit|not_found|extraction|code_error|run_budget_exceeded|budget_exceeded|llm_[a-z0-9_]+)$'
    ReasonCode:
      description: >-
        Code de raison stable (06 § 4.2 : `retried`, `escalated`, `repaired`, `stale`, `reverted`, `cookie_expired`…),
        traduit par la console ; jamais une phrase.
      type: string
      pattern: '^[a-z][a-z0-9_]*$'
    ReasonMessage:
      description: Phrase générée côté serveur, transmise en code et paramètres (traduite par la console, 06 § 4.1).
      type: object
      additionalProperties: false
      required: [code, params]
      properties:
        code:
          $ref: '#/components/schemas/ReasonCode'
        params:
          type: object
          additionalProperties:
            type: [string, number]
...
    Cost:
      description: Coût en dollars ; `null` quand le prix est inconnu (jamais 0 par défaut, 08 § 1).
      type: object
      additionalProperties: false
      required: [llm_usd, proxy_usd, total_usd]
      properties:
        llm_usd:
          type: [number, 'null']
        proxy_usd:
          type: [number, 'null']
        compute_usd:
          type: [number, 'null']
        total_usd:
          type: [number, 'null']
        estimated:
          type: boolean
    Tokens:
      type: object
      additionalProperties: false
      required: [in, cached, out, reasoning, estimated]
      properties:
        in:
          type: integer
        cached:
          type: integer
        out:
          type: integer
        reasoning:
          type: integer
        estimated:
          type: boolean
...
    RunAccepted:
      type: object
      additionalProperties: false
      required: [run_id, state]
      properties:
        run_id:
          type: string
          format: uuid
        state:
          $ref: '#/components/schemas/RunState'
        poll_after_seconds:
          type: [integer, 'null']
...
    RunCancelled:
      type: object
      additionalProperties: false
      required: [run_id, state, cost]
      properties:
        run_id:
          type: string
          format: uuid
        state:
          const: cancelled
        cost:
          $ref: '#/components/schemas/Cost'
    RunSummary:
      description: 'Métadonnées d''un run (état, coût, durée, nombre d''items) : seule vue de l''admin sur les runs d''autrui.'
      type: object
      required: [id, api_id, api_slug, owner_id, trigger, state, outcome, degraded_reasons, failure_class, created_at, started_at, finished_at, duration_ms, cost, items]
      properties:
        id:
          type: string
          format: uuid
        api_id:
          type: string
          format: uuid
        api_slug:
          type: string
        owner_id:
          type: string
          format: uuid
        strategy_version:
          type: [integer, 'null']
        trigger:
          $ref: '#/components/schemas/RunTrigger'
        state:
          $ref: '#/components/schemas/RunState'
        outcome:
          oneOf:
            - $ref: '#/components/schemas/RunOutcome'
            - type: 'null'
        degraded_reasons:
          type: array
          items:
            $ref: '#/components/schemas/ReasonCode'
        failure_class:
          oneOf:
            - $ref: '#/components/schemas/FailureClass'
            - type: 'null'
        retryable:
          type: boolean
        created_at:
          type: string
          format: date-time
        started_at:
          type: [string, 'null']
          format: date-time
        finished_at:
          type: [string, 'null']
          format: date-time
        duration_ms:
          type: [integer, 'null']
        cost:
          $ref: '#/components/schemas/Cost'
        items:
          type: [integer, 'null']
        dataset_id:
          type: [string, 'null']
          format: uuid
        retention_until:
          type: [string, 'null']
          format: date-time
    RunList:
      type: object
      additionalProperties: false
      required: [runs, next_cursor]
      properties:
        runs:
          type: array
          items:
            $ref: '#/components/schemas/RunSummary'
        next_cursor:
          type: [string, 'null']
    RunAttempt:
      description: Un essai de la cascade (exécution × réseau × résultat × coût) ; un essai élagué n'a pas été lancé (INV2).
      type: object
      additionalProperties: false
      required: [index, execution, network, state, est_cost_usd, result, cost_usd, ms]
      properties:
        index:
          type: integer
        execution:
          $ref: '#/components/schemas/Execution'
        network:
          $ref: '#/components/schemas/Network'
        state:
          type: string
          enum: [running, done, pruned]
        pruned_reason:
          type: [string, 'null']
        est_cost_usd:
          type: [number, 'null']
        result:
          description: '`ok`, ou la classe d''échec de l''essai.'
          type: [string, 'null']
        cost_usd:
          type: [number, 'null']
        ms:
          type: [integer, 'null']
        model_id:
          type: [string, 'null']
        prompt_version:
          type: [string, 'null']
        engine:
          type: [string, 'null']
        tokens:
          $ref: '#/components/schemas/Tokens'
        error:
          description: Onglet Erreur (Où, Quoi, Pourquoi, Que faire) en codes stables ; absent pour l'admin sur un run d'autrui.
          oneOf:
            - $ref: '#/components/schemas/ReasonMessage'
            - type: 'null'
    Run:
      description: Détail d'un run (entité Run de 04b § 1) ; `metadata_only` pour l'admin face au run avec session d'autrui.
      allOf:
        - $ref: '#/components/schemas/RunSummary'
        - type: object
          required: [metadata_only, attempts, tokens]
          properties:
            metadata_only:
              type: boolean
            attempts:
              type: array
              items:
                $ref: '#/components/schemas/RunAttempt'
            tokens:
              $ref: '#/components/schemas/Tokens'
            trace_id:
              type: [string, 'null']
            paused_at:
              # Extension de 05 § 4.2 (tâche 3.5, à livrer par 3.1 avec POST /api/runs/{id}/pause) : la console relit la pause
              # à la réouverture d'une enquête, au lieu de la garder seulement en mémoire.
              description: Date de la pause demandée par l'utilisateur (`POST /api/runs/{id}/pause`) ; null hors pause, remis à null par `resume`.
              type: [string, 'null']
              format: date-time
            input:
              description: Entrée du run ; absente pour l'admin sur un run d'autrui.
              type: object
              additionalProperties: true
    RunLogLine:
      type: object
      additionalProperties: false
      required: [seq, at, level, code]
      properties:
        seq:
          type: integer
        at:
          type: string
          format: date-time
        level:
          type: string
          enum: [debug, info, warn, error]
        code:
          type: string
        attempt:
          type: [integer, 'null']
        data:
          type: object
          additionalProperties: true
    RunLogList:
      type: object
      additionalProperties: false
      required: [lines, next_after]
      properties:
        lines:
          type: array
          items:
            $ref: '#/components/schemas/RunLogLine'
        next_after:
          type: [integer, 'null']
...